Ignore pylint error on Travis CI
The error was:

    stbt.py:1680: [E1120(no-value-for-parameter), Display.teardown] No
    value for argument 'self' in unbound method call

referring to `Gst.Event.new_eos()`, which is a static method, but I
guess that pylint doesn't know that.

There are 2 very odd things about this:

1. This error doesn't happen on my own machine running the same version
   of pylint (1.1.0).
2. extra/pylint.conf says `ignored-classes=Event`, so pylint should be
   ignoring any errors related to `Event`.

Note that E1120 exists on pylint 0.28, so users of that version
shouldn't get an error due to the newly-introduced "pylint: disable"
annotation.

Other things I tried before settling on this solution:

* Upgrading Travis CI to pylint 1.2.0 (same error happens; I decided to
  stay on 1.1.0 because that's the version on Ubuntu 14.04).
* Changing extra/pylint.conf to say `ignored-classed=Gst.Event` (same
  error happens).