tests: Run (no of cores +4) tests in parallel
If available `make check` uses GNU parallel to run tests in parallel.  By
default it will run (no of cores) tasks in parallel.  This commit boosts
this number by 4 as on Travis we only have a single core to play with, but
our tests aren't CPU bound so we can further parallelize and further
reduce test run time.

On @drothlis's dual-core laptop the time is reduced from 2m to 1m30s and
on my 4 core laptop it is reduced from 1m2s to 51s.