tests: Fix race between test_stbt_batch_run_killtree and test_killtree
When run in parallel, e.g. with GNU parallel as a part of `make check`
`test_stbt_batch_run_killtree` and `test_killtree` can fail.  You can
reproduce this with

    printf 'test_stbt_batch_run_killtree\ntest_killtree' \
        | parallel ./tests/run-tests.sh -i

The race exists because the `test_stbt_batch_run_killtree` is actually
calling `test_killtree` which in turn tests that it has killed all the
processes that it started up by `grep`ing `ps` output for the program
name - which is the shared by both tests.

This commit fixes this race by naming the program names according to the
pid of the test process (which is the root of the process tree that it's
checking that it's killed).

This race has existed since cb9cb71f5 but we only started seeing it now as
we are trying to get Travis to run the tests in parallel.