Travis: Install GNU parallel to speed up tests
Debian (and by extension Ubuntu) systems have two packages that provide
the command `/usr/bin/parallel`: moreutils and parallel.  Unfortunately
they do not have compatible syntax.  stb-tester depends on moreutils so
we have to install that package.  With GNU parallel (provided by the
Debian package "parallel") we can parallelize the tests we run and so
make them faster.

With this commit we build and install GNU parallel as a part of Travis
commands to allow us to use GNU parallel on Travis' Ubuntu 12.04 and
thus accelerate the testing process.

Reduces Travis build time from ~10mins to 3m47, even though we've added
downloading, compiling and installing GNU parallel to the build process.

Incidentally of the 3m47s when I measured this 2m21s was spent
installing dependencies and 1m26s was spent actually running `make` and
`make check`.