Makefile: don't use parallel --no-notice if this isn't a valid option
With the version of `parallel` I'm using, `--no-notice` is not a valid
option (nor does the message that prints without it occur as per commit
e3aa9fd) and because of that make check(-pylint) fails to run because
`parallel` complains about the option.

This commit modifies the determination of the `parallel` command so that
the `--no-notice` option is not given if it's not valid.