stbt.temporary_x_session: timeout if the x-session fails to start
On my PC, the stbt.test_x11_remote selftest would hang indefinitely
because the file /tmp/.X11-unix/X99 was never being created, and nor
was Xorg return an failed exit code (the `assert` line), thus the function
was caught in an *infinite loop.

This commit doesn't fix the test, but it does add a 10 second timeout
for the x-session to start as expected, which prevents the hang and makes
it easier to test.

*It hung for 2+ minutes, blocking everything else, which AFAIC is infinite