stbt.temporary_x_session: use correct check that process has exited
An object returned by Popen can be checked if it is still running, etc.
However the existing check of looking at the `returncode` attribute would
*always* return `None` because it is only *set* when calling `poll` or
`wait`. [1][2]

This commit changes the check for the value returned by `poll`, which is
None if the command has not exited, other the returncode.

With this change, I can see the reason why this test is failing on my PC
- that is, why I had to implement a timeout in commit ac25d43b - is
because the original command is failing from the outset, i.e. the x-server
is *not* being launched at all.

[1] Example:
In [18]: vmstat = subprocess.Popen(['vmstat', '1'], stdout=open('/dev/null', 'w'))
In [19]: vmstat.returncode
In [20]: vmstat.poll()
In [21]: vmstat.returncode
In [22]: vmstat.terminate()
In [23]: vmstat.returncode
In [24]: vmstat.poll()
Out[24]: -15
In [25]: vmstat.returncode
Out[25]: -15

[2] https://docs.python.org/2/library/subprocess.html#subprocess.Popen.returncode