stbt.uri_to_remote -> factories: [bug] check for empty string not None
Commit 1cd69253 refactored the code for generating remote control objects
into a uri parser (using the `re` module`). However, due to the fact that
in regex a glob '*' can match anything including an empty string, there
was a problem when the default lircd for a lirc remote or the default
hostname (localhost) for a virtual remote were used, because the check

if `param` is None:
     `param` = default

*did not take effect* because `param` is '' (empty string).

This commit fixes this bug by correctly checking for the empty string
instead of None.