stbt-batch report: color uninteresting failures (segv, etc) grey
It's annoying that the *really* uninteresting failures (e.g. sigsegv,
sigterm) are coloured the same as both Precondition and Error type
failures in the batch report, when in reality they are completely
irrelevent to the system-under-test and everything to do with the test
executor. It is also very unhelpful that they are tallied with the
preconditions and errors, because they can skew results, and so I have
added a separate tally for these failures.

Ideally I would like Preconditions and Errors to be listed separately
also. The granularity of "successes", "failures you were looking for"
and "failures you weren't looking for" simply isn't good for quickly
getting an idea of the source of failures.