make check: nosetest more files (all "*.py" files)
37da66f8 tried to make `nosetests` check every python file, but filters
out files with "-" in the name because of the mistaken belief that
`nosetests` can't handle files with "-". In fact, `nosetests` can't
handle files that don't end in ".py"[1], and it was a coincidence that
all our python files that don't end in ".py" also have a "-" in the
filename.

37da66f8 filtered out some files that *do* end in ".py", such as
`stbt-batch.d/report.py` and `tests/validate-ocr.py`.

I had to pass an explicit `--match` pattern to `nosetests` so that it
wouldn't think that `report.testrun` in `stbt-batch.d/report.py` is a
unit test. Similarly, I had to rename `test` in `validate-ocr.py`
because it isn't a unit test.

[1] https://github.com/nose-devs/nose/issues/49