API: OCR: Add stbt.match_text()
`match_text` is intended to be used like `detect_match`, etc, but for
text.  You search for a specific piece of text and it returns extra
information including the bounding box of the detected text.

It uses the tesseract hOCR mode.  hOCR is an XHTML subset with additional
information embedded such as word locations.