gst_hacks: Fix crash on Debian since upgrade to GStreamer 1.4
This fixes a double-free encountered on my Debian system since I upgraded
to GStreamer 1.4, although the potential for the issue has existed longer
than that.

Previously our description of `GstMemoryInfo` as a `ctypes.Structure` only
included the public members mentioned in the documentation, but the
structure in C is actually 32 bytes larger due to private fields and
padding.

We would allocate these structures in Python (via `ctypes`) and then pass
pointers to them to the GStreamer functions `gst_buffer_map` and
`gst_buffer_unmap`.  These functions now make use of these extra bytes but
because we weren't allocating a large enough buffer we were getting a
buffer-overflow causing a double-free to be reported and a crash.

This commit fixes this issue by copying the structure information from the
header file rather than the documentation, which is the true definition of
the structure.