Add `stbt.match` to search a single frame and return MatchResult
Added new API `stbt.match` which was previously conspicuous in its
absence.  It simply searches for a template match in a video frame,
returning a `MatchResult`, and notably not throwing an exception if the
template cannot be found.

This, in combination with the ability to treat `MatchResult` as a
boolean (be15b5d8) makes it much easier to compose matching operations
in user code. e.g.:

    assert stbt.match('template.png') or stbt.match('template2.png')