GStreamer 1.4 support: Work around GStreamer overlay bug
Since upgrading my Debian system to GStreamer 1.4 the test
`test_ocr_on_live_video` has been failing with the error:

    FAIL: test.py: UITestError: Internal data flow error.: Internal data flow error.
    gstbasesrc.c(2933): gst_base_src_loop (): /GstPipeline:pipeline0/GstVideoTestSrc:videotestsrc0:
    streaming task paused, reason not-negotiated (-4)

The source pipeline in this test contains a `textoverlay` element.  This
has changed in GStreamer 1.4 to use the new `GstVideoOverlayComposition`
mechanism optionally allowing sinks to do any required compositing as this
may be more efficient.  This adds additional meta information to the
output caps if `textoverlay` thinks the downstream elements can handle
composition.

In this case downstream can't handle composition, but there is a
`decodebin` in the middle which obscures this from upstream so the meta
caps get added anyway.  When `decodebin` comes to create its output pad
it fails to link with the rest of the downstream elements because the caps
have the meta info and don't match.

This commit works around this bug by making `videotestsrc` output RGB
video that `textoverlay` can composite onto directly.  I have not yet
reported this bug upstream.

For more information see GStreamer bugzilla [#732040]

[#732040]: https://bugzilla.gnome.org/show_bug.cgi?id=732040