stbt.__init__: Stop sink pipeline while source pipeline is restarting
A VidiU capture device restarts its encoder every time video resolution
changes. It results in an EOS message sent to the source bus
that triggers restarting the source pipeline, as implemented in
6b7ba0a3.

That patch is effective with a simple sink pipeline; however, if the
video saving pipeline is active (--save-video option), tests fail
with one of the following errors:
1. GStreamer raises a general stream error: "streaming task paused,
   reason not-negotiated (-4)".
2. If there are multiple subsequent encoder restarts, the second EOS
   message is not sent, therefore the source pipeline never gets
   restarted for the second time, and the test hangs forever or times
   out with NoVideo.
They are not specific to vp8enc; theoraenc and x264enc also produce
the above errors.

If we stop video encoding when the source pipeline sends the EOS
message, and don't restart it until the source pipeline is successfully
restarted, we can avoid the above failures.

Note that it's not enough to set the sink pipeline to STATE_PAUSED.
When we restart video recording, the video encoder raises the error
"negotiation problem ... encoder not initialized". We have to restart
the encoder in order to successfully resume video recording.

Unfortunately the fix has side-effects:
* As the entire sink pipeline is stopped during source pipeline restart,
  the xvimagesink output window is getting closed for this duration.
  As we don't see any video during a pipeline restart anyway, only a
  black or blue screen, it's not a big deal.
* Every time the video encoder restarts it overwrites the video file.
  A solution is provided in a following commit.

Fixes https://github.com/drothlis/stb-tester/issues/211.