stbt.__init__: Create a new video file each time recording is restarted
The commit "Stop sink pipeline while the source pipeline is restarting"
has an issue that filesink overwrites the video file every time it gets
restarted. A possible solution would be to use the filesink in "append"
mode, but appending the next video doesn't recreate the video file
header, and therefore the output file would not be seekable.

A working, though slightly inconvenient solution is to save video to a
new file after every filesink restart. Running

    stbt run --save-video video.webm test.py

with a test that involves 2 source pipeline restarts will create
3 files: video.webm, video1.webm and video2.webm.