stbt.__init__: Save streamable video
This reverts "Create a new video file each time video recording is
restarted" and replaces with an alternative solution.

Save a single streamable file, i.e. no indexes and duration are written.
It's possible to build the video index to make it seekable after the
test is completed with an external application, e.g.

    ffmpeg|avconv -i video.webm -c copy video.seekable.webm