stbt.press: Draw black background behind the key name
`stbt.press` draws the name of the key on the output video. This shows
up on the video recorded by `stbt batch run` for each test run.

When the live TV in the system-under-test is showing white, you couldn't
see the key-name debug text. Now we draw a black box behind the text,
so the debug is always white-on-black.