python packagisation: Rename stbt to stbt.sh, but only in source dir
Rename the "stbt" executable (and the "stbt.in" it's generated from) to
"stbt.sh" (and "stbt.sh.in"). But still install "stbt.sh" as
"$prefix/bin/stbt".

We need to rename it in the source directory so that it doesn't clash
with the python package (added in the next commit) that is a directory
called "stbt".