python packagisation: Move irnetbox.py into stbt package
I considered leaving this outside the stbt package, as it could in
theory be distributed separately to stbt. However my goal with creating
this python package is that I can install it to the system python path,
so that you can import it from anywhere (not just from scripts running
with the "stbt-run" executable, or having to set PYTHONPATH to
$prefix/libexecdir/stbt). But having the stbt package on the system path
is no good if it can't do "import irnetbox". So to simplify
installation, irnetbox.py is going into the stbt package. If someone
wants to pick it up and package it separately, stb-tester's license
allows them to do so.