test_stbt_control_as_stbt_record_control_recorder: Skip if terminal too small
I should figure out how to run the test in some kind of sandboxed
environment where your current terminal size doesn't matter. But for now
this will stop the annoying failures that cause an otherwise-good `make
check` to fail.

This test is currently passing on Travis so the terminal size there is
OK.