python packagisation: stbt.ConfigurationError is Exception not UITestError
This is to avoid a circular import when I move ConfigurationError to the
"stbt.config" sub-package. This is a breaking change.

ConfigurationError inherits from Exception rather than another more
specific exception, so that we have an upgrade path if in future we
decide to change `get_config` to raise `TypeError` and/or `ValueError`.