python packagisation: Remote controls raise RuntimeError, not UITestError
This is to avoid circular imports when I move the control functionality
into a separate sub-package in a future commit.

I want to do away with `UITestError` altogether at some point in the
future; `stbt run` treats anything that isn't a `UITestFailure` as a
test error (see also https://github.com/drothlis/stb-tester/pull/86 ).

Note that `VideoTestSrcControl.press` was raising `UITestFailure` if
the test script specified an invalid key name, but this is really a
test *error*, not a failure.