python packagisation: Move remote controls to stbt.control sub-package
This makes sense because:

* `stbt.control` has no dependencies beyond the standard Python libraries,
  unlike `stbt/__init_.py` which depends on GStreamer, OpenCV, etc.
* The interface between this file and `stbt/__init__.py` is very narrow:
  it's limited to `uri_to_remote` and `uri_to_remote_recorder`.

The original motivation for this (and most of the patch) came from Will
Manley, before I started converting stbt to a python package. Will
wanted to install `fake-lircd` without all the stbt dependencies, for a
light-weight way of providing a proxy for any of these remote controls.
Once we figure out a way of doing `import stbt.control` without loading
all the GStreamer etc. machinery in the main `stbt` package[1], we can
package `fake-lircd` and `stbt/control.py` for deployment.

[1] Perhaps with a multi-directory package that only installs the bits
    we want to use:
    https://docs.python.org/2/tutorial/modules.html#packages-in-multiple-directories