python packagisation: Fix stbt.press pylint warning
Pylint was complaining that "Attribute 'time_of_last_press' defined
outside __init__" in spite of the "pylint:disable=W0201" annotation.
It seems like a bug in pylint where the annotation is ignored if the
object's class is defined in a different file. Storing the time of
the last press as an attribute of the `press` function itself avoids
this warning.