demo-stb: Add .dockerignore to speed docker build
This speeds up the "Sending build context to Docker daemon" step and the
Dockerfile "ADD ." step substantially, when you have a local copy of the
assets (perhaps because you test the demo-stb directly in
chromium-browser on your host system).

On my system this speeds up a no-op `docker build` from 30s to 1s.