test_that_stbt_run_exits_on_ctrl_c: Ignore failures on Travis
Control-C support in stbt is broken on Ubuntu 12.04, so this test fails
intermittently on Travis. See 24c61e01.

I also changed the test to log its periodic "x bottles of beer" messages
directly to the test-run log, not to a separate file, so that when the
test fails you can see what happened.