Smart TV Support: Add command stbt camera
Entry point for stb-tester support for pointing a camera at a TV.  It is
just a placeholder for now.  Real functionality will be added in later
commits.

Building camera support is optional, as is installing it with rpm/deb as
the packages are split into stb-tester and stb-tester-camera.  This will
allow expressing additional dependencies required for camera support but
not for the rest.

stbt camera is experimental for now.  Until we're happy with it you will
have to call it with:

    stbt --with-experimental camera ...

While I expect that the interface will stay roughly the same there are
many details that may change.  This should avoid our users mistaking it
for something completely stable and supported.