report: Better quality of video thumbnail
This increases the size of the thumbnail so that text is readable, but
still keeps a low filesize to load quickly when navigating through lots
of results by pressing the up/down arrow keys (~12k for a screenshot of
some UI elements with a plain background, ~22k with TV in the
background).

This also preserves the aspect ratio of the source video.