tests: Fix run-tests.sh's bash-completion on Ubuntu
The default awk implementation on Ubuntu is mawk, which doesn't like
"()" in the pattern part of the awk program. Escaping the parentheses
with backslashes works with both gawk and mawk.

Oddly mawk doesn't complain about the bare parens in the "-F" pattern,
but I can't escape those for consistency because then gawk complains.