fedora packaging test: Show the manpage for manual inspection
I noticed "ImportError: No module named cv2" when it runs `make doc` as
part of building the RPM, but this easily gets lost in the noise. I'm
not even sure if this will cause a problem in the manpage (seeing as we
commit the changes to README.rst made by `make doc`) so I'm printing the
final manpage to the terminal so you can inspect it.