make check: Don't run integration tests with `PYTHONPATH` set
Since 4ffa6b6 we run our integration tests against an installed version
of stb-tester.  This is great because it is more representative of how
stb-tester will actually be used.  This commit is a further step in this
direction by making PYTHONPATH not point at the source directory, as in
production there is no source directory.

I implemented this while working on Smart TV support as I came across some
problems with imports when using `stbt camera` from the installed
location.  With this change the tests will now catch these issues.