Tests: Skip tests that we don't have the appropriate dependencies for
The goal is to allow ./run-tests -i to work against installed versions
of stb-tester, without requiring any additional dependencies to be
installed beyond stb-tester's existing runtime dependencies.

This should help with automated testing of the packaged versions of
stb-tester as implemented in "extra/fedora/test-packaging.sh", although
further work is required and this script will not yet succeed.