stbt batch report: HTML-escape failure-reason & notes in table
To avoid breaking the table layout.

I'm still not sure whether we should also html-escape these fields in
the individual test-run html page; in one of our projects we put html
links into the notes field, and I can imagine that links in the
failure-reason could be useful too. Maybe the way to go would be to
support a subset of markdown (just bold, italic, and links)?