make check-nosetests: Ignore user's PYTHONPATH
I had PYTHONPATH pointing at a different, installed, version of stbt and
it caused nose's test-collector to fail.

We already set PYTHONPATH for `make check-pylint`. Since 68983e0 we
neither set nor clear PYTHONPATH for `tests/run-tests.sh` (aka
`make check-integrationtests`) but I'm not sure whether I should leave
that as it is, or set `PYTHONPATH=` -- so I'll leave it as it is.