stbt config: Allow specifying multiple config files in $STBT_CONFIG_FILE
Convert $STBT_CONFIG_FILE to a list similar to $PATH and $PYTHONPATH
where the item specified at the beginning takes precedence.

Example:
    STBT_CONFIG_FILE=/path/to/temporary.conf:/path/to/permanent.conf
Parameters are searched for in temporary.conf first and permanent.conf
next.

Our use-case of this feature is to split the test specific config to a
set of permanent parameters that we version control, and a set of
temporary parameters that we are changing all the time:

- `source_pipeline` and `sink_pipeline` are specific to the video
  capture hardware that we hardly ever change. We would find it useful
  to track changes in capture device set-up, therefore we would like to
  version control these parameters and fetch them from a repository
  for every test run. They would be read from `permanent.conf` in the
  example above. They cannot go into the user config because we have
  multiple capture devices assigned to the same user.

- `control` depends on the model of the set-top-box, therefore it is
  frequently modified, and we would allow testers to change it as they
  like. It would go into `temporary.conf` in the example.