stbt run: Write tracing information to STBT_TRACING_SOCKET
With this pull request what stbt is doing right now is written to a socket
given by the `STBT_TRACING_SOCKET` environment variable.  This allows the
current state to be communicated to the user through an interactive UI,
making it easier to understand what stbt is doing at any one time.  We
also write these tracing logs to disk.

The state that is tracked includes which test is currently executing, what
line is currently executing, etc.  This could be hooked up to a live UI
and should be useful during test replay when investigating what has gone
wrong.

Imagine a description of the state of stbt in a data structure like:

    {
        "test_run": {
            "current_line": {
                "file": "tests/my_file.py",
                "line": 123,
            },
            "test": {
                "name": "tests/my_file.py::test_that_this_rocks",
                "file": "tests/my_file.py",
                "function": "test_that_this_rocks,
                "line": 87
            }
        }
    }

This data structure is ephermal and only exists while the test is
executing.  Instead of sending the whole data structure over the socket
every time there is a change we instead serialise changes to subtrees of
this heirarical data structure.  These changes are sent over the socket
The serialisation includes timestamps such that the state of the system
at any time in the past can be reconstructed, potentially useful for
replaying old test results.

The change serialisation format is \\r\\n seperated JSON dictionaries.
Here's an example entry pretty printed for clarity's sake (in reality it
would be on a single line):

    {
        "state_change": {
            "time": {"$date": "2014-11-28T20:55:26.092343Z"},
            "changes": {
                "test_job.current_line": {
                    "file": "tests/my_file.py",
                    "line": 654
                }
            }
        }
    }

This record indicates that the current line changed to line 654 of file
"tests/my_file.py" at 20:55:26.0992343 UTC on 2014-11-28.  Notes:

* The root object has a single key "state_change".  This allows extensions
  in the future for other types of messages to be sent over the same
  protocol.

* The state_change message has two keys:

    * "time" which is the time of the change using mongodb style date
      format `{"$date": "2014-11-28T20:55:26.092343Z"}`.  The date
      timestamp is in ISO8601 format.  The fact that this is an object
      rather than just a string may seem a little odd but JSON has no
      native datetime type so this is necessary to specify the type.

    * "changes" is a dictionary of changes made to the state of the
      system.  The keys identify the value that is changing, and the
      values are the new value of the respective key.  The key is dot
      seperated strings identifying the subtree of the heirarical data
      structure that should be replaced.  e.g:

          "test_run.current_line": {"cow": "moo"}

      means:

          date["test_run"]["current_line"] = {"cow": "moo"}

      The fact that multiple values may be replaced in one message allows
      atomic changes to happen to the hierarchy.