Makefile: Don't update API docs in README.rst during packaging
Summary: Stb-tester core developers now need to run `make
update-api-docs` to update README.rst after changing or adding public
API docstrings.

--

stbt's Python API docs in README.rst are checked in to the repository;
they need to be updated (and committed) whenever public APIs change or
new ones are added. The man page is generated from README.rst.

Previously README.rst was updated automatically when running `make` or
`make install`. This can cause problems when building the packages for
Fedora or Ubuntu. Because the source python file's timestamp might be
newer than README.rst (even if README.rst was fully up to date --
perhaps the most recent change to the python source file didn't affect
any public APIs), running `make` (as part of the packaging process)
would run the recipe to update README.rst, which runs pydoc, which
imports the stbt python files, which fails if run-time dependencies
aren't installed. The end result is that the API documentation
disappears from the man page.

We want to be able to build the packages without installing stbt's
run-time dependencies. Also the `api-doc.sh` script is very fragile
(I want to replace it with something like Sphinx at some point) and it
doesn't report failures, so I don't want it to be run automatically in
places where no one is checking the result.

With this change, the man page produced by the Ubuntu Launchpad and
Fedora COPR binary package build infrastructure will exactly match the
README.rst that we provide as input to that build process.

Updating the README.rst is the responsibility of the stb-tester
developers (as it always was). The difference is that now you have to
run `make update-api-docs` instead of `make doc`.

To make sure we don't forget, I've added a self-test that ensures `make
update-api-docs` doesn't change README.rst. This new test covers
everything tested by the existing selftests in `test-make-doc.sh` so
I've removed them.